Re #13540 Moved test for parseKeyValueString()
