+++ Code/Mantid/MantidQt/CustomInterfaces/test/ReflMainViewPresenterTest.h
@@ -707,33 +707,6 @@ public:
     AnalysisDataService::Instance().remove("TestWorkspace");
   }
 
-  void testParseKeyValueString() {
-    std::map<std::string, std::string> kvp =
-        ReflMainViewPresenter::parseKeyValueString(
-            "a = 1,b=2.0, c=3, d='1,2,3',e=\"4,5,6\",f=1+1=2, g = '\\''");
-
-    TS_ASSERT_EQUALS(kvp["a"], "1");
-    TS_ASSERT_EQUALS(kvp["b"], "2.0");
-    TS_ASSERT_EQUALS(kvp["c"], "3");
-    TS_ASSERT_EQUALS(kvp["d"], "1,2,3");
-    TS_ASSERT_EQUALS(kvp["e"], "4,5,6");
-    TS_ASSERT_EQUALS(kvp["f"], "1+1=2");
-    TS_ASSERT_EQUALS(kvp["g"], "'");
-
-    TS_ASSERT_THROWS(
-        ReflMainViewPresenter::parseKeyValueString("a = 1, b = 2, c = 3,"),
-        std::runtime_error);
-    TS_ASSERT_THROWS(
-        ReflMainViewPresenter::parseKeyValueString("a = 1, b = 2, c = 3,d"),
-        std::runtime_error);
-    TS_ASSERT_THROWS(ReflMainViewPresenter::parseKeyValueString(",a = 1"),
-                     std::runtime_error);
-    TS_ASSERT_THROWS(ReflMainViewPresenter::parseKeyValueString(",a = 1 = 2,="),
-                     std::runtime_error);
-    TS_ASSERT_THROWS(ReflMainViewPresenter::parseKeyValueString("=,=,="),
-                     std::runtime_error);
-  }
-
   void testPromptSaveAfterAppendRow() {
     MockView mockView;
     ReflMainViewPresenter presenter(&mockView);
